+++ src/Components/Header/Header.js
@@ -1,0 +1,17 @@
+import React from 'react';
+// C S S
+import '/.Header.css';
+
+class Header extends React.Component {
+  render() {
+    return (
+      <Header>
+        <h1>
+          City Explorer
+        </h1>
+      </Header>
+    )
+  }
+}
+
+export default Header;
